Move the dispatcher frame to the controller class
